vehiculos 100%

+++ Solid_Cronometro/Clases/RNVehiculo.cs
@@ -5,8 +5,8 @@ namespace Solid_Cronometro.Clases
         internal static readonly byte nro_ruedas_2 = 2;
         internal static readonly byte nro_ruedas_4 = 4;
         internal static readonly byte nro_ruedas_6 = 6;
-        internal static readonly int min_presion_rueda = 20;
-        internal static readonly int max_presion_rueda = 60;
+        internal static readonly float min_presion_rueda = 20;
+        internal static readonly float max_presion_rueda = 60;
         internal static readonly int long_placa = 6;
         internal static readonly int long_mod_min = 2;
         internal static readonly int long_marca_min = 2;
